code review: Origin Footer 컴포넌트 분리
