+++ src/components/Footer/OriginFooter/style.ts
